requirements.txt
aniso8601==8.0.0
appdirs==1.4.4
astroid==2.4.1
beautifulsoup4==4.6.0
bootstrap4==0.1.0
boto==2.49.0
boto3==1.9.188
botocore==1.12.188
certifi==2019.6.16
cfe==0.0.15
chardet==3.0.4
click==7.1.2
colorama==0.4.1
cycler==0.10.0
diff-match-patch==20181111
distlib==0.3.0
dj-database-url==0.5.0
Django==2.2
django-appconf==1.0.3
django-authtools==1.7.0
django-bootstrap4==0.0.8
django-braces==1.13.0
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-environ==0.4.5
django-filebrowser==3.11.1
django-formtools==2.1
django-grappelli==2.12.3
django-heroku==0.3.1
django-import-export==1.2.0
django-js-asset==1.2.2
django-reversion==3.0.4
django-social-share==1.3.2
django-storages==1.7.1
djangorestframework==3.10.0
docutils==0.14
easy-thumbnails==2.6
et-xmlfile==1.0.1
filelock==3.0.12
Flask==1.1.1
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.7
future==0.17.1
gunicorn==19.9.0
h5py==2.10.0
httplib2==0.9.2
idna==2.8
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mysqlclient==1.4.2.post1
numpy==1.18.4
oauth2client==4.1.3
oauthlib==3.0.2
opencv-python==4.2.0.34
openpyxl==2.6.2
pandas==1.0.4
Pillow==6.1.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
pylint==2.5.2
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
rsa==4.0
s3transfer==0.2.1
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.9.0
six==1.12.0
sqlparse==0.3.0
tablib==2.0.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.1
tqdm==4.46.0
transfer==0.40
typed-ast==1.4.1
urllib3==1.25.3
virtualenv==20.0.21
Werkzeug==0.15.4
whitenoise==4.1.3
wrapt==1.12.1
zipp==3.1.0
